@@ -11,12 +11,14 @@
   "homepage": "https://github.com/homebridge/HAP-NodeJS",
   "license": "Apache-2.0",
   "scripts": {
-    "clean": "rimraf dist/",
-    "build": "rimraf dist/ && tsc && node .github/workflows/node-persist-ignore.js",
+    "clean": "rimraf dist && rimraf coverage",
+    "build": "rimraf dist && tsc && node .github/workflows/node-persist-ignore.js",
     "prepublishOnly": "npm run build",
     "postpublish": "npm run clean",
     "test": "jest",
-    "start": "node dist/BridgedCore.js"
+    "test-coverage": "jest --coverage",
+    "start": "node dist/BridgedCore.js",
+    "docs": "typedoc src/index.ts"
   },
   "keywords": [
     "hap-nodejs",
@@ -49,24 +51,30 @@
   ],
   "dependencies": {
     "@homebridge/ciao": "~1.1.2",
+    "bonjour-hap": "~3.6.2",
+    "debug": "^4.3.1",
     "fast-srp-hap": "2.0.2",
-    "tweetnacl": "^1.0.3",
-    "debug": "^4.1.1",
-    "decimal.js": "^10.2.0",
+    "futoin-hkdf": "~1.3.2",
     "ip": "^1.1.3",
     "node-persist": "^0.0.11",
-    "futoin-hkdf": "~1.3.2"
+    "source-map-support": "^0.5.19",
+    "tslib": "^2.1.0",
+    "tweetnacl": "^1.0.3"
   },
   "devDependencies": {
     "@types/debug": "^4.1.5",
-    "@types/jest": "^26.0.13",
-    "@types/node": "^10.17.20",
-    "jest": "^26.4.2",
+    "@types/escape-html": "^1.0.0",
+    "@types/jest": "^26.0.20",
+    "@types/node": "^10.17.50",
+    "commander": "^6.2.1",
+    "escape-html": "^1.0.3",
+    "jest": "^26.6.3",
     "rimraf": "^3.0.2",
-    "semver": "^7.3.2",
-    "simple-plist": "^1.1.0",
-    "ts-jest": "^26.3.0",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.2"
+    "semver": "^7.3.4",
+    "simple-plist": "^1.1.1",
+    "ts-jest": "^26.4.4",
+    "ts-node": "^9.1.1",
+    "typedoc": "0.20.14",
+    "typescript": "^4.1.3"
   }
 }
