@@ -4,23 +4,20 @@ import {
   AccessoryEventTypes,
   Categories,
   Characteristic,
-  CharacteristicEventTypes, CharacteristicSetCallback,
   CharacteristicValue,
-  NodeCallback,
   Service,
   uuid,
   VoidCallback
 } from '..';
 
-var FAKE_FAN: Record<string, any> = {
+const FAKE_FAN: Record<string, any> = {
   powerOn: false,
   rSpeed: 100,
   setPowerOn: (on: CharacteristicValue) => {
-    if(on){
+    if (on) {
       //put your code here to turn on the fan
       FAKE_FAN.powerOn = on;
-    }
-    else{
+    } else {
       //put your code here to turn off the fan
       FAKE_FAN.powerOn = on;
     }
@@ -34,10 +31,10 @@ var FAKE_FAN: Record<string, any> = {
     //put your code here to identify the fan
     console.log("Fan Identified!");
   }
-}
+};
 
 // This is the Accessory that we'll return to HAP-NodeJS that represents our fake fan.
-var fan = exports.accessory = new Accessory('Fan', uuid.generate('hap-nodejs:accessories:Fan'));
+const fan = exports.accessory = new Accessory('Fan', uuid.generate('hap-nodejs:accessories:Fan'));
 
 // Add properties for publishing (in case we're using Core.js and not BridgedCore.js)
 // @ts-ignore
@@ -59,13 +56,11 @@ fan.on(AccessoryEventTypes.IDENTIFY, (paired: boolean, callback: VoidCallback) =
 });
 
 // Add the actual Fan Service and listen for change events from iOS.
-// We can see the complete list of Services and Characteristics in `lib/gen/HomeKit.ts`
 fan
   .addService(Service.Fan, "Fan") // services exposed to the user should have "names" like "Fake Light" for us
   .getCharacteristic(Characteristic.On)!
-  .on(CharacteristicEventTypes.SET, (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
+  .onSet((value) => {
     FAKE_FAN.setPowerOn(value);
-    callback(); // Our fake Fan is synchronous - this value has been successfully set
   });
 
 // We want to intercept requests for our current power state so we can query the hardware itself instead of
@@ -73,19 +68,16 @@ fan
 fan
   .getService(Service.Fan)!
   .getCharacteristic(Characteristic.On)!
-  .on(CharacteristicEventTypes.GET, (callback: NodeCallback<CharacteristicValue>) => {
+  .onGet(() => {
 
     // this event is emitted when you ask Siri directly whether your fan is on or not. you might query
     // the fan hardware itself to find this out, then call the callback. But if you take longer than a
     // few seconds to respond, Siri will give up.
 
-    var err = null; // in case there were any problems
-
     if (FAKE_FAN.powerOn) {
-      callback(err, true);
-    }
-    else {
-      callback(err, false);
+      return true;
+    } else {
+      return false;
     }
   });
 
@@ -93,10 +85,9 @@ fan
 fan
   .getService(Service.Fan)!
   .addCharacteristic(Characteristic.RotationSpeed)
-  .on(CharacteristicEventTypes.GET, (callback: NodeCallback<CharacteristicValue>) => {
-    callback(null, FAKE_FAN.rSpeed);
+  .onGet(async () => {
+    return FAKE_FAN.rSpeed;
   })
-  .on(CharacteristicEventTypes.SET, (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
+  .onSet(async (value) => {
     FAKE_FAN.setSpeed(value);
-    callback();
   })
