@@ -3,8 +3,9 @@ import {
   AccessoryEventTypes,
   Categories,
   Characteristic,
-  CharacteristicEventTypes, CharacteristicSetCallback,
-  CharacteristicValue,
+  CharacteristicEventTypes,
+  CharacteristicSetCallback,
+  CharacteristicValue, ColorUtils,
   NodeCallback,
   Service,
   uuid,
@@ -25,7 +26,7 @@ class LightControllerClass {
   hue: CharacteristicValue = 0; //current hue
   saturation: CharacteristicValue = 0; //current saturation
 
-  outputLogs = false; //output logs
+  outputLogs = true; //output logs
 
   setPower(status: CharacteristicValue) { //set power of accessory
     if(this.outputLogs) console.log("Turning the '%s' %s", this.name, status ? "on" : "off");
@@ -77,10 +78,10 @@ const LightController = new LightControllerClass();
 // Generate a consistent UUID for our light Accessory that will remain the same even when
 // restarting our server. We use the `uuid.generate` helper function to create a deterministic
 // UUID based on an arbitrary "namespace" and the word "light".
-var lightUUID = uuid.generate('hap-nodejs:accessories:light' + LightController.name);
+const lightUUID = uuid.generate('hap-nodejs:accessories:light' + LightController.name);
 
 // This is the Accessory that we'll return to HAP-NodeJS that represents our light.
-var lightAccessory = exports.accessory = new Accessory(LightController.name as string, lightUUID);
+const lightAccessory = exports.accessory = new Accessory(LightController.name as string, lightUUID);
 
 // Add properties for publishing (in case we're using Core.js and not BridgedCore.js)
 // @ts-ignore
@@ -103,11 +104,9 @@ lightAccessory.on(AccessoryEventTypes.IDENTIFY, (paired: boolean, callback: Void
   callback();
 });
 
-// Add the actual Lightbulb Service and listen for change events from iOS.
-// We can see the complete list of Services and Characteristics in `lib/gen/HomeKit.ts`
-lightAccessory
-  .addService(Service.Lightbulb, LightController.name) // services exposed to the user should have "names" like "Light" for this case
-  .getCharacteristic(Characteristic.On)!
+const lightbulb = lightAccessory.addService(Service.Lightbulb, LightController.name); // services exposed to the user should have "names" like "Light" for this case
+
+lightbulb.getCharacteristic(Characteristic.On)
   .on(CharacteristicEventTypes.SET, (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
     LightController.setPower(value);
 
@@ -133,9 +132,7 @@ lightAccessory
 //   .updateValue(true);
 
 // also add an "optional" Characteristic for Brightness
-lightAccessory
-  .getService(Service.Lightbulb)!
-  .addCharacteristic(Characteristic.Brightness)
+lightbulb.addCharacteristic(Characteristic.Brightness)
   .on(CharacteristicEventTypes.SET, (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
     LightController.setBrightness(value);
     callback();
@@ -144,10 +141,9 @@ lightAccessory
     callback(null, LightController.getBrightness());
   });
 
+
 // also add an "optional" Characteristic for Saturation
-lightAccessory
-  .getService(Service.Lightbulb)!
-  .addCharacteristic(Characteristic.Saturation)!
+lightbulb.addCharacteristic(Characteristic.Saturation)
   .on(CharacteristicEventTypes.SET, (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
     LightController.setSaturation(value);
     callback();
@@ -157,9 +153,7 @@ lightAccessory
   });
 
 // also add an "optional" Characteristic for Hue
-lightAccessory
-  .getService(Service.Lightbulb)!
-  .addCharacteristic(Characteristic.Hue)
+lightbulb.addCharacteristic(Characteristic.Hue)
   .on(CharacteristicEventTypes.SET, (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
     LightController.setHue(value);
     callback();
