@@ -14,32 +14,36 @@ import { Nullable } from '../types';
 let err: Nullable<Error> = null; // in case there were any problems
 
 // here's a fake hardware device that we'll expose to HomeKit
-var FAKE_OUTLET = {
+const FAKE_OUTLET = {
   powerOn: false,
-    setPowerOn: (on: CharacteristicValue) => {
+  setPowerOn: (on: CharacteristicValue) => {
     console.log("Turning the outlet %s!...", on ? "on" : "off");
     if (on) {
-          FAKE_OUTLET.powerOn = true;
-          if(err) { return console.log(err); }
-          console.log("...outlet is now on.");
+      FAKE_OUTLET.powerOn = true;
+      if (err) {
+        return console.log(err);
+      }
+      console.log("...outlet is now on.");
     } else {
-          FAKE_OUTLET.powerOn = false;
-          if(err) { return console.log(err); }
-          console.log("...outlet is now off.");
+      FAKE_OUTLET.powerOn = false;
+      if (err) {
+        return console.log(err);
+      }
+      console.log("...outlet is now off.");
     }
   },
-    identify: function() {
+  identify: function () {
     console.log("Identify the outlet.");
-    }
-}
+  }
+};
 
 // Generate a consistent UUID for our outlet Accessory that will remain the same even when
 // restarting our server. We use the `uuid.generate` helper function to create a deterministic
 // UUID based on an arbitrary "namespace" and the accessory name.
-var outletUUID = uuid.generate('hap-nodejs:accessories:Outlet');
+const outletUUID = uuid.generate('hap-nodejs:accessories:Outlet');
 
 // This is the Accessory that we'll return to HAP-NodeJS that represents our fake light.
-var outlet = exports.accessory = new Accessory('Outlet', outletUUID);
+const outlet = exports.accessory = new Accessory('Outlet', outletUUID);
 
 // Add properties for publishing (in case we're using Core.js and not BridgedCore.js)
 // @ts-ignore
@@ -63,7 +67,6 @@ outlet.on(AccessoryEventTypes.IDENTIFY, function(paired: boolean, callback: Void
 });
 
 // Add the actual outlet Service and listen for change events from iOS.
-// We can see the complete list of Services and Characteristics in `lib/gen/HomeKit.ts`
 outlet
   .addService(Service.Outlet, "Fake Outlet") // services exposed to the user should have "names" like "Fake Light" for us
   .getCharacteristic(Characteristic.On)!
@@ -83,7 +86,7 @@ outlet
     // the light hardware itself to find this out, then call the callback. But if you take longer than a
     // few seconds to respond, Siri will give up.
 
-    var err = null; // in case there were any problems
+    const err = null; // in case there were any problems
 
     if (FAKE_OUTLET.powerOn) {
       console.log("Are we on? Yes.");
