@@ -1,26 +1,26 @@
 // here's a fake temperature sensor device that we'll expose to HomeKit
 import { Accessory, Categories, Characteristic, CharacteristicEventTypes, CharacteristicValue, NodeCallback, Service, uuid } from '..';
 
-var FAKE_SENSOR = {
+const FAKE_SENSOR = {
   currentTemperature: 50,
-  getTemperature: function() {
+  getTemperature: function () {
     console.log("Getting the current temperature!");
     return FAKE_SENSOR.currentTemperature;
   },
-  randomizeTemperature: function() {
+  randomizeTemperature: function () {
     // randomize temperature to a value between 0 and 100
     FAKE_SENSOR.currentTemperature = Math.round(Math.random() * 100);
   }
-}
+};
 
 
 // Generate a consistent UUID for our Temperature Sensor Accessory that will remain the same
 // even when restarting our server. We use the `uuid.generate` helper function to create
 // a deterministic UUID based on an arbitrary "namespace" and the string "temperature-sensor".
-var sensorUUID = uuid.generate('hap-nodejs:accessories:temperature-sensor');
+const sensorUUID = uuid.generate('hap-nodejs:accessories:temperature-sensor');
 
 // This is the Accessory that we'll return to HAP-NodeJS that represents our fake lock.
-var sensor = exports.accessory = new Accessory('Temperature Sensor', sensorUUID);
+const sensor = exports.accessory = new Accessory('Temperature Sensor', sensorUUID);
 
 // Add properties for publishing (in case we're using Core.js and not BridgedCore.js)
 // @ts-ignore
@@ -31,7 +31,6 @@ sensor.pincode = "031-45-154";
 sensor.category = Categories.SENSOR;
 
 // Add the actual TemperatureSensor Service.
-// We can see the complete list of Services and Characteristics in `lib/gen/HomeKit.ts`
 sensor
   .addService(Service.TemperatureSensor)!
   .getCharacteristic(Characteristic.CurrentTemperature)!
