@@ -1,11 +1,11 @@
-import './gen';
 import {
-  Accessory,
+  Accessory, AccessoryEventTypes, Bridge,
   Categories,
   Characteristic,
   CharacteristicEventTypes,
-  Controller, ControllerServiceMap,
-  ControllerType,
+  Controller,
+  ControllerIdentifier,
+  ControllerServiceMap,
   Service,
   uuid
 } from '..';
@@ -13,7 +13,9 @@ import {
 
 class TestController implements Controller {
 
-  readonly controllerType: ControllerType = "test-type";
+  controllerId(): ControllerIdentifier {
+    return "test-id";
+  }
 
   constructServices(): ControllerServiceMap {
     const lightService = new Service.Lightbulb('', '');
@@ -35,6 +37,9 @@ class TestController implements Controller {
   }
 
   configureServices(): void {}
+
+  handleControllerRemoved(): void {
+  }
 
 }
 
@@ -70,6 +75,45 @@ describe('Accessory', () => {
       expect(() => {
         new Accessory('Test', 'test');
       }).toThrow('not a valid UUID');
+    });
+  });
+
+  describe("characteristicWarning", () => {
+    it("should emit characteristic warning", () => {
+      let accessory = new Accessory("Test Accessory", uuid.generate("Test"));
+      let handler = jest.fn();
+      accessory.on(AccessoryEventTypes.CHARACTERISTIC_WARNING, handler);
+
+      let service = accessory.addService(Service.Lightbulb, "Light");
+      let on = service.getCharacteristic(Characteristic.On);
+
+      on.updateValue({});
+      expect(handler).toHaveBeenCalledTimes(1)
+    });
+
+    it("should forward characteristic on bridged accessory", () => {
+      let bridge = new Bridge("Test bridge", uuid.generate("bridge test"));
+
+      let accessory = new Accessory("Test Accessory", uuid.generate("Test"));
+      bridge.addBridgedAccessory(accessory);
+
+      let handler = jest.fn();
+      bridge.on(AccessoryEventTypes.CHARACTERISTIC_WARNING, handler);
+      accessory.on(AccessoryEventTypes.CHARACTERISTIC_WARNING, handler);
+
+      let service = accessory.addService(Service.Lightbulb, "Light");
+      let on = service.getCharacteristic(Characteristic.On);
+
+      on.updateValue({});
+      expect(handler).toHaveBeenCalledTimes(2)
+    });
+
+    it("should run without characteristic warning handler", () => {
+      let accessory = new Accessory("Test Accessory", uuid.generate("Test"));
+      let service = accessory.addService(Service.Lightbulb, "Light");
+      let on = service.getCharacteristic(Characteristic.On);
+
+      on.updateValue({});
     });
   });
 
