@@ -1,20 +1,19 @@
-import createDebug from 'debug';
 import assert from 'assert';
+import crypto from 'crypto';
+import createDebug from 'debug';
+import { EventEmitter, EventEmitter as NodeEventEmitter } from "events";
+import net, { Socket } from 'net';
+import { HAPConnection, HAPConnectionEvent } from "../util/eventedhttp";
 
 import * as hapCrypto from '../util/hapCrypto';
-import {DataStreamParser, DataStreamReader, DataStreamWriter, Int64} from './DataStreamParser';
-import crypto from 'crypto';
-import net, {Socket} from 'net';
-import {HAPSessionEvents, Session} from "../util/eventedhttp";
-import {EventEmitter as NodeEventEmitter} from "events";
-import {EventEmitter} from "../EventEmitter";
+import { DataStreamParser, DataStreamReader, DataStreamWriter, Int64 } from './DataStreamParser';
 import Timeout = NodeJS.Timeout;
 
 const debug = createDebug('HAP-NodeJS:DataStream:Server');
 
 export type PreparedDataStreamSession = {
 
-    session: Session, // reference to the hap session which created the request
+    connection: HAPConnection, // reference to the hap session which created the request
 
     accessoryToControllerEncryptionKey: Buffer,
     controllerToAccessoryEncryptionKey: Buffer,
@@ -26,6 +25,7 @@ export type PreparedDataStreamSession = {
 
 }
 
+export type PrepareSessionCallback = (error?: Error, preparedSession?: PreparedDataStreamSession) => void;
 export type EventHandler = (message: Record<any, any>) => void;
 export type RequestHandler = (id: number, message: Record<any, any>) => void;
 export type ResponseHandler = (error: Error | undefined, status: HDSStatus | undefined, message: Record<any, any>) => void;
@@ -60,6 +60,7 @@ export const enum Topics { // a collection of currently known topics grouped by 
 }
 
 export enum HDSStatus {
+    // noinspection JSUnusedGlobalSymbols
     SUCCESS = 0,
     OUT_OF_MEMORY = 1,
     TIMEOUT = 2,
@@ -70,6 +71,7 @@ export enum HDSStatus {
 }
 
 export enum DataSendCloseReason { // close reason used in the dataSend protocol
+    // noinspection JSUnusedGlobalSymbols
     NORMAL = 0,
     NOT_ALLOWED = 1,
     BUSY = 2,
@@ -121,27 +123,30 @@ type DataStreamMessage = {
     message: Record<any, any>,
 }
 
-export const enum DataStreamServerEvents {
+export const enum DataStreamServerEvent {
+    /**
+     * This event is emitted when a new client socket is received. At this point we have no idea to what
+     * hap session this connection will be matched.
+     */
     CONNECTION_OPENED = "connection-opened",
+    /**
+     * This event is emitted when the socket of a connection gets closed.
+     */
     CONNECTION_CLOSED = "connection-closed",
 }
 
-export type DataStreamServerEventMap = {
-    [DataStreamServerEvents.CONNECTION_OPENED]: (connection: DataStreamConnection) => void;
-    [DataStreamServerEvents.CONNECTION_CLOSED]: (connection: DataStreamConnection) => void;
+export declare interface DataStreamServer {
+    on(event: "connection-opened", listener: (connection: DataStreamConnection) => void): this;
+    on(event: "connection-closed", listener: (connection: DataStreamConnection) => void): this;
+
+    emit(event: "connection-opened", connection: DataStreamConnection): boolean;
+    emit(event: "connection-closed", connection: DataStreamConnection): boolean;
 }
 
 /**
  * DataStreamServer which listens for incoming tcp connections and handles identification of new connections
- *
- * @event 'connection-opened': (connection: DataStreamConnection) => void
- *        This event is emitted when a new client socket is received. At this point we have no idea to what
- *        hap session this connection will be matched.
- *
- * @event 'connection-closed': (connection: DataStreamConnection) => void
- *        This event is emitted when the socket of a connection gets closed.
  */
-export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
+export class DataStreamServer extends EventEmitter {
 
     static readonly version = "1.0";
 
@@ -153,15 +158,14 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
     private tcpServer?: net.Server;
     private tcpPort?: number;
 
-    preparedSessions: PreparedDataStreamSession[];
-    private connections: DataStreamConnection[];
+    preparedSessions: PreparedDataStreamSession[] = [];
+    private readonly connections: DataStreamConnection[] = [];
+    private removeListenersOnceClosed = false;
 
     private readonly internalEventEmitter: NodeEventEmitter = new NodeEventEmitter(); // used for message event and message request handlers
 
-    constructor() {
+    public constructor() {
         super();
-        this.preparedSessions = [];
-        this.connections = [];
     }
 
     /**
@@ -173,7 +177,7 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
      * @param event {string | Topics} - name of the event (also referred to as topic. See {Topics} for some known ones)
      * @param handler {GlobalEventHandler} - function to be called for every occurring event
      */
-    onEventMessage(protocol: string | Protocols, event: string | Topics, handler: GlobalEventHandler): this {
+    public onEventMessage(protocol: string | Protocols, event: string | Topics, handler: GlobalEventHandler): this {
         this.internalEventEmitter.on(protocol + "-e-" + event, handler);
         return this;
     }
@@ -185,7 +189,7 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
      * @param event {string | Topics} - name of the event (also referred to as topic. See {Topics} for some known ones)
      * @param handler {GlobalEventHandler} - registered event handler
      */
-    removeEventHandler(protocol: string | Protocols, event: string | Topics, handler: GlobalEventHandler): this {
+    public removeEventHandler(protocol: string | Protocols, event: string | Topics, handler: GlobalEventHandler): this {
         this.internalEventEmitter.removeListener(protocol + "-e-" + event, handler);
         return this;
     }
@@ -199,7 +203,7 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
      * @param request {string | Topics} - name of the request (also referred to as topic. See {Topics} for some known ones)
      * @param handler {GlobalRequestHandler} - function to be called for every occurring request
      */
-    onRequestMessage(protocol: string | Protocols, request: string | Topics, handler: GlobalRequestHandler): this {
+    public onRequestMessage(protocol: string | Protocols, request: string | Topics, handler: GlobalRequestHandler): this {
         this.internalEventEmitter.on(protocol + "-r-" + request, handler);
         return this;
     }
@@ -211,33 +215,40 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
      * @param request {string | Topics} - name of the request (also referred to as topic. See {Topics} for some known ones)
      * @param handler {GlobalRequestHandler} - registered request handler
      */
-    removeRequestHandler(protocol: string | Protocols, request: string | Topics, handler: GlobalRequestHandler): this {
+    public removeRequestHandler(protocol: string | Protocols, request: string | Topics, handler: GlobalRequestHandler): this {
         this.internalEventEmitter.removeListener(protocol + "-r-" + request, handler);
         return this;
     }
 
-    prepareSession(session: Session, controllerKeySalt: Buffer, callback: (preparedSession: PreparedDataStreamSession) => void) {
-        debug("Preparing for incoming HDS connection from session %s", session.sessionID);
+    public prepareSession(connection: HAPConnection, controllerKeySalt: Buffer, callback: PrepareSessionCallback) {
+        debug("Preparing for incoming HDS connection from %s", connection.sessionID);
         const accessoryKeySalt = crypto.randomBytes(32);
         const salt = Buffer.concat([controllerKeySalt, accessoryKeySalt]);
 
-        const accessoryToControllerEncryptionKey = hapCrypto.HKDF("sha512", salt, session.encryption!.sharedSec, DataStreamServer.accessoryToControllerInfo, 32);
-        const controllerToAccessoryEncryptionKey = hapCrypto.HKDF("sha512", salt, session.encryption!.sharedSec, DataStreamServer.controllerToAccessoryInfo, 32);
+        const accessoryToControllerEncryptionKey = hapCrypto.HKDF("sha512", salt, connection.encryption!.sharedSecret, DataStreamServer.accessoryToControllerInfo, 32);
+        const controllerToAccessoryEncryptionKey = hapCrypto.HKDF("sha512", salt, connection.encryption!.sharedSecret, DataStreamServer.controllerToAccessoryInfo, 32);
 
         const preparedSession: PreparedDataStreamSession = {
-            session: session,
+            connection: connection,
             accessoryToControllerEncryptionKey: accessoryToControllerEncryptionKey,
             controllerToAccessoryEncryptionKey: controllerToAccessoryEncryptionKey,
             accessoryKeySalt: accessoryKeySalt,
             connectTimeout: setTimeout(() => this.timeoutPreparedSession(preparedSession), 10000),
         };
+        preparedSession.connectTimeout!.unref();
         this.preparedSessions.push(preparedSession);
 
-        this.checkTCPServerEstablished(preparedSession, () => callback(preparedSession));
+        this.checkTCPServerEstablished(preparedSession, (error) => {
+            if (error) {
+                callback(error);
+            } else {
+                callback(undefined, preparedSession);
+            }
+        });
     }
 
     private timeoutPreparedSession(preparedSession: PreparedDataStreamSession) {
-        debug("Prepared HDS session timed out out since no connection was opened for 10 seconds (%s)", preparedSession.session.sessionID);
+        debug("Prepared HDS session timed out out since no connection was opened for 10 seconds (%s)", preparedSession.connection.sessionID);
         const index = this.preparedSessions.indexOf(preparedSession);
         if (index >= 0) {
             this.preparedSessions.splice(index, 1);
@@ -246,7 +257,7 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
         this.checkCloseable();
     }
 
-    private checkTCPServerEstablished(preparedSession: PreparedDataStreamSession, callback: () => void) {
+    private checkTCPServerEstablished(preparedSession: PreparedDataStreamSession, callback: (error?: Error) => void) {
         switch (this.state) {
             case ServerState.UNINITIALIZED:
                 debug("Starting up TCP server.");
@@ -275,7 +286,7 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
         }
     }
 
-    private listening(preparedSession: PreparedDataStreamSession, callback: () => void) {
+    private listening(preparedSession: PreparedDataStreamSession, callback: (error?: Error) => void) {
         this.state = ServerState.LISTENING;
 
         const address = this.tcpServer!.address();
@@ -292,13 +303,13 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
         debug("[%s] New DataStream connection was established", socket.remoteAddress);
         const connection = new DataStreamConnection(socket);
 
-        connection.on(DataStreamConnectionEvents.IDENTIFICATION, this.handleSessionIdentification.bind(this, connection));
-        connection.on(DataStreamConnectionEvents.HANDLE_MESSAGE_GLOBALLY, this.handleMessageGlobally.bind(this, connection));
-        connection.on(DataStreamConnectionEvents.CLOSED, this.connectionClosed.bind(this, connection));
+        connection.on(DataStreamConnectionEvent.IDENTIFICATION, this.handleSessionIdentification.bind(this, connection));
+        connection.on(DataStreamConnectionEvent.HANDLE_MESSAGE_GLOBALLY, this.handleMessageGlobally.bind(this, connection));
+        connection.on(DataStreamConnectionEvent.CLOSED, this.connectionClosed.bind(this, connection));
 
         this.connections.push(connection);
 
-        this.emit(DataStreamServerEvents.CONNECTION_OPENED, connection);
+        this.emit(DataStreamServerEvent.CONNECTION_OPENED, connection);
     }
 
     private handleSessionIdentification(connection: DataStreamConnection, firstFrame: HDSFrame, callback: IdentificationCallback) {
@@ -316,7 +327,7 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
         callback(identifiedSession);
 
         if (identifiedSession) {
-            debug("[%s] Connection was successfully identified (linked with sessionId: %s)", connection._remoteAddress, identifiedSession.session.sessionID);
+            debug("[%s] Connection was successfully identified (linked with sessionId: %s)", connection.remoteAddress, identifiedSession.connection.sessionID);
             const index = this.preparedSessions.indexOf(identifiedSession);
             if (index >= 0) {
                 this.preparedSessions.splice(index, 1);
@@ -327,9 +338,9 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
 
             // we have currently no experience with data stream connections, maybe it would be good to index active connections
             // by their hap sessionId in order to clear out old but still open connections when the controller opens a new one
-            // on the other han the keepAlive should handle that also :thinking:
+            // on the other hand the keepAlive should handle that also :thinking:
         } else { // we looped through all session and didn't find anything
-            debug("[%s] Could not identify connection. Terminating.", connection._remoteAddress);
+            debug("[%s] Could not identify connection. Terminating.", connection.remoteAddress);
             connection.close(); // disconnecting since first message was not a valid hello
         }
     }
@@ -352,32 +363,52 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
             hadListeners = this.internalEventEmitter.emit(message.protocol + separator + message.topic, connection, ...args);
         } catch (error) {
             hadListeners = true;
-            debug("[%s] Error occurred while dispatching handler for HDS message: %o", connection._remoteAddress, message);
+            debug("[%s] Error occurred while dispatching handler for HDS message: %o", connection.remoteAddress, message);
             debug(error.stack);
         }
 
         if (!hadListeners) {
-            debug("[%s] WARNING no handler was found for message: %o", connection._remoteAddress, message);
+            debug("[%s] WARNING no handler was found for message: %o", connection.remoteAddress, message);
         }
     }
 
     private connectionClosed(connection: DataStreamConnection) {
-        debug("[%s] DataStream connection closed", connection._remoteAddress);
+        debug("[%s] DataStream connection closed", connection.remoteAddress);
 
         this.connections.splice(this.connections.indexOf(connection), 1);
-        this.emit(DataStreamServerEvents.CONNECTION_CLOSED, connection);
+        this.emit(DataStreamServerEvent.CONNECTION_CLOSED, connection);
 
         this.checkCloseable();
+
+        if (this.state === ServerState.CLOSING && this.removeListenersOnceClosed && this.connections.length === 0) {
+            this.removeAllListeners(); // see this.destroy()
+        }
     }
 
     private checkCloseable() {
-        if (this.connections.length === 0 && this.preparedSessions.length === 0) {
+        if (this.connections.length === 0 && this.preparedSessions.length === 0 && this.state < ServerState.CLOSING) {
             debug("Last connection disconnected. Closing the server now.");
 
             this.state = ServerState.CLOSING;
-            // noinspection JSIgnoredPromiseFromCall
             this.tcpServer!.close();
         }
+    }
+
+    /**
+     * This method will fully stop the DataStreamServer
+     */
+    public destroy(): void {
+        if (this.state > ServerState.UNINITIALIZED && this.state < ServerState.CLOSING) {
+            this.tcpServer!.close();
+            for (const connection of this.connections) {
+                connection.close();
+            }
+        }
+
+        this.state = ServerState.CLOSING;
+
+        this.removeListenersOnceClosed = true;
+        this.internalEventEmitter.removeAllListeners();
     }
 
     private closed() {
@@ -389,48 +420,52 @@ export class DataStreamServer extends EventEmitter<DataStreamServerEventMap> {
 
 }
 
-export const enum DataStreamConnectionEvents {
+export type IdentificationCallback = (identifiedSession?: PreparedDataStreamSession) => void;
+
+export const enum DataStreamConnectionEvent {
+    /**
+     * This event is emitted when the first HDSFrame is received from a new connection.
+     * The connection expects the handler to identify the connection by trying to match the decryption keys.
+     * If identification was successful the PreparedDataStreamSession should be supplied to the callback,
+     * otherwise undefined should be supplied.
+     */
     IDENTIFICATION = "identification",
+    /**
+     * This event is emitted when no handler could be found for the given protocol of a event or request message.
+     */
     HANDLE_MESSAGE_GLOBALLY = "handle-message-globally",
+    /**
+     * This event is emitted when the socket of the connection was closed.
+     */
     CLOSED = "closed",
 }
 
-export type IdentificationCallback = (identifiedSession?: PreparedDataStreamSession) => void;
+export declare interface DataStreamConnection {
+    on(event: "identification", listener: (frame: HDSFrame, callback: IdentificationCallback) => void): this;
+    on(event: "handle-message-globally", listener: (message: DataStreamMessage) => void): this;
+    on(event: "closed", listener: () => void): this;
 
-export type DataStreamConnectionEventMap = {
-    [DataStreamConnectionEvents.IDENTIFICATION]: (frame: HDSFrame, callback: IdentificationCallback) => void;
-    [DataStreamConnectionEvents.HANDLE_MESSAGE_GLOBALLY]: (message: DataStreamMessage) => void;
-    [DataStreamConnectionEvents.CLOSED]: () => void;
+    emit(event: "identification", frame: HDSFrame, callback: IdentificationCallback): boolean;
+    emit(event: "handle-message-globally", message: DataStreamMessage): boolean;
+    emit(event: "closed"): boolean;
 }
 
 /**
  * DataStream connection which holds any necessary state information, encryption an decryption keys, manages
  * protocol handlers and also handles sending and receiving of data stream frames.
- *
- * @event 'identification': (frame: HDSFrame, callback: IdentificationCallback) => void
- *        This event is emitted when the first HDSFrame is received from a new connection.
- *        The connection expects the handler to identify the connection by trying to match the decryption keys.
- *        If identification was successful the PreparedDataStreamSession should be supplied to the callback,
- *        otherwise undefined should be supplied.
- *
- * @event 'handle-message-globally': (message: DataStreamMessage) => void
- *        This event is emitted when no handler could be found for the given protocol of a event or request message.
- *
- * @event 'closed': () => void
- *        This event is emitted when the socket of the connection was closed.
  */
-export class DataStreamConnection extends EventEmitter<DataStreamConnectionEventMap> {
+export class DataStreamConnection extends EventEmitter {
 
     private static readonly MAX_PAYLOAD_LENGTH = 0b11111111111111111111;
 
     private socket: Socket;
-    private session?: Session; // reference to the hap session. is present when state > UNIDENTIFIED
-    readonly _remoteAddress: string;
+    private connection?: HAPConnection; // reference to the hap connection. is present when state > UNIDENTIFIED
+    readonly remoteAddress: string;
     /*
         Since our DataStream server does only listen on one port and this port is supplied to every client
         which wants to connect, we do not really know which client is who when we receive a tcp connection.
         Thus, we find the correct PreparedDataStreamSession object by testing the encryption keys of all available
-        prepared sessions. Then we can reference this connection with the correct session and mark it as identified.
+        prepared sessions. Then we can reference this hds connection with the correct hap connection and mark it as identified.
      */
     private state: ConnectionState = ConnectionState.UNIDENTIFIED;
 
@@ -444,6 +479,7 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
 
     private frameBuffer?: Buffer; // used to store incomplete HDS frames
 
+    private readonly hapConnectionClosedListener: () => void;
     private protocolHandlers: Record<string, DataStreamProtocolHandler> = {}; // used to store protocolHandlers identified by their protocol name
 
     private responseHandlers: Record<number, ResponseHandler> = {}; // used to store responseHandlers indexed by their respective requestId
@@ -454,7 +490,7 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
     constructor(socket: Socket) {
         super();
         this.socket = socket;
-        this._remoteAddress = socket.remoteAddress!;
+        this.remoteAddress = socket.remoteAddress!;
 
         this.socket.setNoDelay(true); // disable Nagle algorithm
         this.socket.setKeepAlive(true);
@@ -464,6 +500,8 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
         this.controllerToAccessoryNonce = 0;
         this.controllerToAccessoryNonceBuffer = Buffer.alloc(8);
 
+        this.hapConnectionClosedListener = this.onHAPSessionClosed.bind(this);
+
         this.addProtocolHandler(Protocols.CONTROL, {
            requestHandler: {
                [Topics.HELLO]: this.handleHello.bind(this)
@@ -471,18 +509,18 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
         });
 
         this.helloTimer = setTimeout(() => {
-            debug("[%s] Hello message did not arrive in time. Killing the connection", this._remoteAddress);
+            debug("[%s] Hello message did not arrive in time. Killing the connection", this.remoteAddress);
             this.close();
         }, 10000);
 
         this.socket.on('data', this.onSocketData.bind(this));
         this.socket.on('error', this.onSocketError.bind(this));
-        this.socket.on('close', this.onSocketClose.bind(this)); // we MUST register for this event, otherwise the error will bubble up to the top and crash the node process entirely.
+        this.socket.on('close', this.onSocketClose.bind(this));
     }
 
     private handleHello(id: number, _message: Record<any, any>) {
         // that hello is indeed the _first_ message received is verified in onSocketData(...)
-        debug("[%s] Received hello message from client", this._remoteAddress);
+        debug("[%s] Received hello message from client", this.remoteAddress);
 
         clearTimeout(this.helloTimer!);
         this.helloTimer = undefined;
@@ -608,27 +646,28 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
         if (this.state === ConnectionState.UNIDENTIFIED) {
             // at the beginning we are only interested in trying to decrypt the first frame in order to test decryption keys
             const firstFrame = frames[frameIndex++];
-            this.emit(DataStreamConnectionEvents.IDENTIFICATION, firstFrame, (identifiedSession?: PreparedDataStreamSession) => {
+            this.emit(DataStreamConnectionEvent.IDENTIFICATION, firstFrame, (identifiedSession?: PreparedDataStreamSession) => {
                if (identifiedSession) {
-                   // horray, we found our session
-                   this.session = identifiedSession.session;
+                   // horray, we found our connection
+                   this.connection = identifiedSession.connection;
                    this.accessoryToControllerEncryptionKey = identifiedSession.accessoryToControllerEncryptionKey;
                    this.controllerToAccessoryEncryptionKey = identifiedSession.controllerToAccessoryEncryptionKey;
                    this.state = ConnectionState.EXPECTING_HELLO;
 
-                   this.session.on(HAPSessionEvents.CLOSED, this.onHAPSessionClosed.bind(this)); // register close listener
+                   // below listener is removed in .close()
+                   this.connection.on(HAPConnectionEvent.CLOSED, this.hapConnectionClosedListener); // register close listener
                }
             });
 
             if (this.state === ConnectionState.UNIDENTIFIED) {
-                // did not find a prepared session, server already closed this connection; nothing to do here
+                // did not find a prepared connection, server already closed this connection; nothing to do here
                 return;
             }
         }
 
         for (; frameIndex < frames.length; frameIndex++) { // decrypt all remaining frames
             if (!this.decryptHDSFrame(frames[frameIndex])) {
-                debug("[%s] HDS frame decryption or authentication failed. Connection will be terminated!", this._remoteAddress);
+                debug("[%s] HDS frame decryption or authentication failed. Connection will be terminated!", this.remoteAddress);
                 this.close();
                 return;
             }
@@ -641,7 +680,7 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
 
             if (firstMessage.protocol !== Protocols.CONTROL || firstMessage.type !== MessageType.REQUEST || firstMessage.topic !== Topics.HELLO) {
                 // first message is not the expected hello request
-                debug("[%s] First message received was not the expected hello message. Instead got: %o", this._remoteAddress, firstMessage);
+                debug("[%s] First message received was not the expected hello message. Instead got: %o", this.remoteAddress, firstMessage);
                 this.close();
                 return;
             }
@@ -668,7 +707,7 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
                 try {
                     responseHandler(undefined, message.status!, message.message);
                 } catch (error) {
-                    debug("[%s] Error occurred while dispatching response handler for HDS message: %o", this._remoteAddress, message);
+                    debug("[%s] Error occurred while dispatching response handler for HDS message: %o", this.remoteAddress, message);
                     debug(error.stack);
                 }
                 delete this.responseHandlers[message.id!];
@@ -676,38 +715,38 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
                 const handler = this.protocolHandlers[message.protocol];
                 if (handler === undefined) {
                     // send message to the server to check if there are some global handlers for it
-                    this.emit(DataStreamConnectionEvents.HANDLE_MESSAGE_GLOBALLY, message);
+                    this.emit(DataStreamConnectionEvent.HANDLE_MESSAGE_GLOBALLY, message);
                     return;
                 }
 
                 if (message.type === MessageType.EVENT) {
                     let eventHandler: EventHandler;
                     if (!handler.eventHandler || !(eventHandler = handler.eventHandler[message.topic])) {
-                        debug("[%s] WARNING no event handler was found for message: %o", this._remoteAddress, message);
+                        debug("[%s] WARNING no event handler was found for message: %o", this.remoteAddress, message);
                         return;
                     }
 
                     try {
                         eventHandler(message.message);
                     } catch (error) {
-                        debug("[%s] Error occurred while dispatching event handler for HDS message: %o", this._remoteAddress, message);
+                        debug("[%s] Error occurred while dispatching event handler for HDS message: %o", this.remoteAddress, message);
                         debug(error.stack);
                     }
                 } else if (message.type === MessageType.REQUEST) {
                     let requestHandler: RequestHandler;
                     if (!handler.requestHandler || !(requestHandler = handler.requestHandler[message.topic])) {
-                        debug("[%s] WARNING no request handler was found for message: %o", this._remoteAddress, message);
+                        debug("[%s] WARNING no request handler was found for message: %o", this.remoteAddress, message);
                         return;
                     }
 
                     try {
                         requestHandler(message.id!, message.message);
                     } catch (error) {
-                        debug("[%s] Error occurred while dispatching request handler for HDS message: %o", this._remoteAddress, message);
+                        debug("[%s] Error occurred while dispatching request handler for HDS message: %o", this.remoteAddress, message);
                         debug(error.stack);
                     }
                 } else {
-                    debug("[%s] Encountered unknown message type with id %d", this._remoteAddress, message.type);
+                    debug("[%s] Encountered unknown message type with id %d", this.remoteAddress, message.type);
                 }
             }
         })
@@ -733,7 +772,7 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
             const payloadLength = data.readUIntBE(frameBegin + 1, 3); // read 24-bit big-endian uint length field
 
             if (payloadLength > DataStreamConnection.MAX_PAYLOAD_LENGTH) {
-                debug("[%s] Connection send payload with size bigger than the maximum allow for data stream", this._remoteAddress);
+                debug("[%s] Connection send payload with size bigger than the maximum allow for data stream", this.remoteAddress);
                 this.close();
                 return [];
             }
@@ -764,7 +803,7 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
                 };
                 frames.push(hdsFrame);
             } else {
-                debug("[%s] Encountered unknown payload type %d for payload: %s", this._remoteAddress, plaintextPayload.toString('hex'));
+                debug("[%s] Encountered unknown payload type %d for payload: %s", this.remoteAddress, plaintextPayload.toString('hex'));
             }
         }
 
@@ -810,7 +849,7 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
                 headerDictionary = DataStreamParser.decode(headerPayload);
                 headerPayload.finished();
             } catch (error) {
-                debug("[%s] Failed to decode header payload: %s", this._remoteAddress, error.message);
+                debug("[%s] Failed to decode header payload: %s", this.remoteAddress, error.message);
                 return;
             }
 
@@ -818,7 +857,7 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
                 messageDictionary = DataStreamParser.decode(messagePayload);
                 messagePayload.finished();
             } catch (error) {
-                debug("[%s] Failed to decode message payload: %s (header: %o)", this._remoteAddress, error.message, headerDictionary);
+                debug("[%s] Failed to decode message payload: %s (header: %o)", this.remoteAddress, error.message, headerDictionary);
                 return;
             }
 
@@ -841,7 +880,7 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
                 id = headerDictionary["id"];
                 status = headerDictionary["status"];
             } else {
-                debug("[%s] Encountered unknown payload header format: %o (message: %o)", this._remoteAddress, headerDictionary, messageDictionary);
+                debug("[%s] Encountered unknown payload header format: %o (message: %o)", this.remoteAddress, headerDictionary, messageDictionary);
                 return;
             }
 
@@ -909,24 +948,28 @@ export class DataStreamConnection extends EventEmitter<DataStreamConnectionEvent
             return; // connection is already closing/closed
         }
 
+        this.connection?.removeListener(HAPConnectionEvent.CLOSED, this.hapConnectionClosedListener);
+
         this.state = ConnectionState.CLOSING;
         this.socket.end();
     }
 
     private onHAPSessionClosed() {
-        // If the hap session is closed it is probably also a good idea to close the data stream session
-        debug("[%s] HAP session disconnected. Also closing DataStream connection now.", this._remoteAddress);
+        // If the hap connection is closed it is probably also a good idea to close the data stream connection
+        debug("[%s] HAP connection disconnected. Also closing DataStream connection now.", this.remoteAddress);
         this.close();
     }
 
     private onSocketError(error: Error) {
-        debug("[%s] Encountered socket error: %s", this._remoteAddress, error.message);
+        debug("[%s] Encountered socket error: %s", this.remoteAddress, error.message);
         // onSocketClose will be called next
     }
 
     private onSocketClose() {
+        // this instance is now considered completely dead
         this.state = ConnectionState.CLOSED;
-        this.emit(DataStreamConnectionEvents.CLOSED);
+        this.emit(DataStreamConnectionEvent.CLOSED);
+        this.removeAllListeners();
     }
 
 }
