@@ -4,15 +4,17 @@
  */
 export function clone<T, U>(object: T, extend?: U): T & U {
 
-  var cloned = {} as Record<any, any>;
+  const cloned = {} as Record<any, any>;
 
-  for (var key in object) {
-    cloned[key] = object[key];
+  for (const [ key, value ] of Object.entries(object)) {
+    cloned[key] = value;
   }
 
-  for (var key2 in extend) {
-    cloned[key2] = extend[key2];
+  if (extend) {
+    for (const [ key, value ] of Object.entries(extend)) {
+      cloned[key] = value;
+    }
   }
 
   return cloned;
-};
+}
